// packages/rpc-collector/src/data.generated.ts
export interface RawRpcRecord {
  readonly url: string;
  readonly tracking: string;
  readonly isOpenSource?: boolean;
}